[audio] Fix 32-bit overflow in AudioStreamInfo duration conversions
`frames_to_microseconds()` computes `frames * 1000000` in uint32 arithmetic, which
overflows at 4295 frames -- 97.4 ms of audio at 44.1 kHz, 89.5 ms at 48 kHz. The
result comes back short by exactly 2^32 / sample_rate per wrap (97391.5 us at
44.1 kHz), silently and with no diagnostic. One second of 44.1 kHz audio reports as
26085 us; 4295 frames reports as 1 us.

No current caller reaches this. All three pass sub-buffer frame counts
(SPDIF_BLOCK_SAMPLES, frames_zeroed, silence_frames), so this is a latent defect
rather than an observed regression. But nothing in the signature says a caller may
not exceed it: the parameter is a uint32_t frame count, and a caller reporting a
buffer's occupancy rather than one chunk of it passes 4295 frames routinely -- a
100 ms ring at 44.1 kHz is 4410.

Fixed with a 64-bit intermediate. The same 32-bit product appears in
frames_to_milliseconds_with_remainder(), ms_to_frames(), ms_to_samples() and
ms_to_bytes(); those need tens of seconds in a single call to reach it, so they are
more remote still, but the arithmetic is identical and is fixed alongside.

Adds tests/components/audio/, which had no C++ tests. Six of the eight fail against
the unfixed conversion, including a monotonicity check that catches the boundary
without hard-coding it -- a longer queue must never report a shorter duration.

+++ esphome/components/audio/audio.cpp
@@ -23,14 +23,25 @@ AudioStreamInfo::AudioStreamInfo(uint8_t bits_per_sample, uint8_t channels, uint
 }
 
 uint32_t AudioStreamInfo::frames_to_microseconds(uint32_t frames) const {
-  return (frames * 1000000 + (this->sample_rate_ >> 1)) / this->sample_rate_;
+  // 64-bit intermediate: `frames * 1000000` in uint32 arithmetic overflows at 4295 frames, which is
+  // 97.4 ms of audio at 44.1 kHz and 89.5 ms at 48 kHz. The wrapped result is short by exactly
+  // 2^32 / sample_rate per wrap (97391.5 us at 44.1 kHz), silently and with no diagnostic.
+  //
+  // Every current caller passes a sub-buffer frame count and stays well under that, but nothing in
+  // the signature says it must: the parameter is a uint32_t frame count, and a caller reporting a
+  // buffer's occupancy rather than one chunk of it exceeds 4295 frames routinely.
+  return static_cast<uint32_t>((static_cast<uint64_t>(frames) * 1000000 + (this->sample_rate_ >> 1)) /
+                               this->sample_rate_);
 }
 
 uint32_t AudioStreamInfo::frames_to_milliseconds_with_remainder(uint32_t *total_frames) const {
   uint32_t unprocessable_frames = *total_frames % (this->sample_rate_ / this->ms_sample_rate_gcd_);
   uint32_t frames_for_ms_calculation = *total_frames - unprocessable_frames;
 
-  uint32_t playback_ms = (frames_for_ms_calculation * 1000) / this->sample_rate_;
+  // 64-bit for the same reason as frames_to_microseconds(); this one only wraps past ~97 s of audio
+  // in one call, so it is a latent case rather than an observed one, but it is the same defect.
+  uint32_t playback_ms =
+      static_cast<uint32_t>((static_cast<uint64_t>(frames_for_ms_calculation) * 1000) / this->sample_rate_);
   *total_frames = unprocessable_frames;
   return playback_ms;
 }
